Moved some possibly relevant code into the new Biolink3 object.
